changes to cluster

+++ cluster/pipeline_executable.py
@@ -15,6 +15,7 @@ import pandas as pd
 import pathlib
 import warnings
 import shutil
+import zipfile
 import os
 warnings.filterwarnings("ignore")
 #os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
@@ -24,7 +25,8 @@ warnings.filterwarnings("ignore")
 ######################################
 ## User passed arguments
 remote_folder = sys.argv[1]                             # Path to the remote Folder
-remote_folder.replace('\\', '/')                        # Converting backslash to forward slash if needed 
+#remote_folder.replace('\\', '/')                        # Converting backslash to forward slash if needed 
+data_folder_path = pathlib.Path(remote_folder)
 send_data_to_NAS = int(sys.argv[2])                     # Flag to send data back to NAS
 diamter_nucleus = int(sys.argv[3])                      # approximate nucleus size in pixels
 diameter_cytosol = int(sys.argv[4])  #250               # approximate cytosol size in pixels
@@ -33,8 +35,30 @@ psf_yx_1 = int(sys.argv[6])      #150                   # Theoretical size of th
 nucleus_channel= int(sys.argv[7])                       # Channel to pass to python for nucleus segmentation
 cyto_channel= int(sys.argv[8])                          # Channel to pass to python for cytosol segmentation
 FISH_channel= int(sys.argv[9])                          # Channel to pass to python for spot detection
-FISH_second_channel= int(sys.argv[10])                  # Channel to pass to python for spot detection
+
+
+FISH_second_channel= sys.argv[10]                  # Channel to pass to python for spot detection
+if FISH_second_channel == 'None':
+  FISH_second_channel = None
+else:
+  FISH_second_channel= int(FISH_second_channel)
+
+  
 output_name = sys.argv[11]                              # Output file name
+# Path to credentials
+path_to_config_file = pathlib.Path(sys.argv[12])
+
+connect_to_NAS= int(sys.argv[13])
+
+path_to_masks_dir= sys.argv[14]
+if path_to_masks_dir == 'None':
+  path_to_masks_dir = None
+else:
+  path_to_masks_dir = pathlib.Path(path_to_masks_dir )
+  
+optimization_segmentation_method= pathlib.Path(sys.argv[15])
+if optimization_segmentation_method == 'None':
+  optimization_segmentation_method = None
 
 
 # Deffining directories
@@ -47,43 +71,65 @@ import fish_analyses as fa
 
 # Printing banner
 fa.Banner().print_banner()
-
-# Path to credentials
-#try:
-path_to_config_file = pathlib.Path(sys.argv[12])
-#path_to_config_file = current_dir.parents[0].joinpath('config.yml')
-#desktop_path = pathlib.Path.home()/'Desktop'
-#path_to_config_file = desktop_path.joinpath('config.yml')
-  
-  
-#except:
-#desktop_path = pathlib.Path.home()/'Desktop'
-#path_to_config_file = desktop_path.joinpath('config.yml')
-  
 share_name = 'share'
 
-remote_folder_path = pathlib.Path(remote_folder)
-
-name_final_folder = (remote_folder_path.name +'___nuc_' + str(diamter_nucleus) +
+name_final_folder = (data_folder_path.name +'___nuc_' + str(diamter_nucleus) +
                 '__cyto_' + str(diameter_cytosol) +
                 '__psfz_' + str(psf_z_1) +
                 '__psfyx_' + str(psf_yx_1) )
 
+def download_data_NAS(path_to_config_file,data_folder_path, path_to_masks_dir,share_name,timeout=200):
+    # Downloading data from NAS
+    local_folder_path = pathlib.Path().absolute().joinpath('temp_' + data_folder_path.name)
+    fa.NASConnection(path_to_config_file,share_name = share_name).copy_files(data_folder_path, local_folder_path,timeout=timeout)
+    local_data_dir = local_folder_path     # path to a folder with images.
+    # Downloading masks from NAS
+    if not (path_to_masks_dir is None):
+        local_folder_path_masks = pathlib.Path().absolute().joinpath('temp_masks_' + data_folder_path.name )
+        zip_file_path = local_folder_path_masks.joinpath( path_to_masks_dir.stem +'.zip')
+        fa.NASConnection(path_to_config_file,share_name = share_name).download_file(path_to_masks_dir, local_folder_path_masks,timeout=timeout)
+        # Unzip downloaded images and update mask directory
+        file_to_unzip = zipfile.ZipFile(str(zip_file_path)) # opens zip
+        # Iterates for each file in zip file
+        for file_in_zip in file_to_unzip.namelist():
+            # Extracts data to specific folder
+            file_to_unzip.extract(file_in_zip,local_folder_path_masks)
+        # Closes the zip file
+        file_to_unzip.close()
+        # removes the original zip file
+        os.remove(zip_file_path)
+        masks_dir = local_folder_path_masks
+    else:
+        masks_dir = None
+    return local_data_dir, masks_dir
+
 # Download data from NAS
-local_folder_path = pathlib.Path().absolute().joinpath('temp_' + remote_folder_path.name)
-fa.NASConnection(path_to_config_file,share_name = share_name).copy_files(remote_folder_path, local_folder_path,timeout=120)
+if connect_to_NAS == True:
+    share_name = 'share'
+    local_data_dir, masks_dir= download_data_NAS(path_to_config_file,data_folder_path, path_to_masks_dir,share_name,timeout=200)
+else:
+    local_data_dir = data_folder_path 
+    masks_dir = path_to_masks_dir 
+
 
 # Parameters for the code
-data_dir = local_folder_path     # path to a folder with images.
+#data_dir = local_folder_path     # path to a folder with images.
 channels_with_cytosol = [nucleus_channel,cyto_channel]            # list or int indicating the channels where the cytosol is detectable
 channels_with_nucleus = nucleus_channel                # list or int indicating the channels where the nucleus is detectable
 
 # Deffining FISH Channels
-if FISH_second_channel==0:
+if (FISH_second_channel==0) or (FISH_second_channel==None):
   channels_with_FISH = [FISH_channel]               # list or int with the channels with FISH spots that are used for the quantification
 else:
   channels_with_FISH = [FISH_channel,FISH_second_channel ]
-  
+
+
+# Detecting if images need to be merged
+is_needed_to_merge_images = fa.MergeChannels(local_data_dir, substring_to_detect_in_file_name = '.*_C0.tif', save_figure =1).checking_images()
+if is_needed_to_merge_images == True:
+    list_file_names, list_images, number_images, output_to_path = fa.MergeChannels(local_data_dir, substring_to_detect_in_file_name = '.*_C0.tif', save_figure =1).merge()
+    local_data_dir = local_data_dir.joinpath('merged')
+
 # Parameters for FISH detection
 voxel_size_z = 500                       # Microscope conversion px to nanometers in the z axis.
 voxel_size_yx = 103                      # Microscope conversion px to nanometers in the xy axis.
@@ -94,14 +140,9 @@ list_psfs = [ [psf_z_1, psf_yx_1] ]
 minimum_spots_cluster = 2                # The number of spots in a neighborhood for a point to be considered as a core point (from which a cluster is expanded). This includes the point itself.
 show_plots=True                          # Flag to display plots
 
-# Detecting if images need to be merged
-is_needed_to_merge_images = fa.MergeChannels(data_dir, substring_to_detect_in_file_name = '.*_C0.tif', save_figure =1).checking_images()
-if is_needed_to_merge_images == True:
-    list_file_names, list_images, number_images, output_to_path = fa.MergeChannels(data_dir, substring_to_detect_in_file_name = '.*_C0.tif', save_figure =1).merge()
-    data_dir = data_dir.joinpath('merged')
-
 # Running the pipeline
-dataframe_FISH,_,_,_ = fa.PipelineFISH(data_dir, channels_with_cytosol, channels_with_nucleus, channels_with_FISH,diamter_nucleus, diameter_cytosol, minimum_spots_cluster,list_voxels=list_voxels, list_psfs=list_psfs ,show_plot=show_plots,file_name_str=remote_folder_path.name).run()
+#dataframe_FISH,_,_,_ = fa.PipelineFISH(data_dir, channels_with_cytosol, channels_with_nucleus, channels_with_FISH,diamter_nucleus, diameter_cytosol, minimum_spots_cluster,list_voxels=list_voxels, list_psfs=list_psfs ,show_plot=show_plots,file_name_str=remote_folder_path.name).run()
+dataframe_FISH,_,_,_ = fa.PipelineFISH(local_data_dir, channels_with_cytosol, channels_with_nucleus, channels_with_FISH,diamter_nucleus, diameter_cytosol, minimum_spots_cluster, masks_dir=masks_dir,  list_voxels=list_voxels, list_psfs=list_psfs, show_plot=show_plots, file_name_str =data_folder_path.name, optimization_segmentation_method = optimization_segmentation_method ).run()
 
 # Number of cells
 spot_type_selected = 0
@@ -115,8 +156,6 @@ number_of_spots_per_cell_cytosol = [len( dataframe_FISH.loc[  (dataframe_FISH['c
 number_of_spots_per_cell_nucleus = [len( dataframe_FISH.loc[  (dataframe_FISH['cell_id']==i) &  (dataframe_FISH['is_cluster']==False) & (dataframe_FISH['is_nuc']==True) & (dataframe_FISH['spot_type']==spot_type_selected)    ].spot_id) for i in range(0, number_cells)]
 # Number of TS per cell.
 number_of_TS_per_cell = [len( dataframe_FISH.loc[  (dataframe_FISH['cell_id']==i) &  (dataframe_FISH['is_cluster']==True) & (dataframe_FISH['is_nuc']==True) & (dataframe_FISH['spot_type']==spot_type_selected) & (dataframe_FISH['cluster_size'] >=4) ].spot_id) for i in range(0, number_cells)]
-#number_of_TS_per_cell= np.asarray(number_of_TS_per_cell)
-#number_of_TS_per_cell=number_of_TS_per_cell[number_of_TS_per_cell>0]   
 # Number of RNA in a TS
 ts_size =  dataframe_FISH.loc[   (dataframe_FISH['is_cluster']==True) & (dataframe_FISH['is_nuc']==True)  & (dataframe_FISH['spot_type']==spot_type_selected)   ].cluster_size.values
 # Size of each cell
@@ -125,7 +164,7 @@ cell_size = dataframe_FISH.loc[  (dataframe_FISH['spot_id']==0)  ].cell_area_px.
 # Plotting intensity distributions
 plt.style.use('ggplot')  # ggplot  #default
 def plot_probability_distribution(data_to_plot, numBins = 10, title='', xlab='', ylab='', color='r', subplots=False, show_grid=True, fig=plt.figure() ):
-  n, bins, patches = plt.hist(data_to_plot,bins=numBins,density=False,color=color)
+  n, bins, _ = plt.hist(data_to_plot,bins=numBins,density=False,color=color)
   plt.xlabel(xlab, size=16)
   plt.ylabel(ylab, size=16)
   plt.grid(show_grid)
@@ -149,47 +188,68 @@ f.add_subplot(1,5,4)
 plot_probability_distribution(ts_size, numBins=20,    title='Clusters in nucleus', xlab='RNA per Cluster', ylab=ylab, fig=f, color='orangered')
 f.add_subplot(1,5,5) 
 plot_probability_distribution(number_of_TS_per_cell ,  numBins=20, title='Number TS per cell', xlab='[TS (>= 4 rna)]', ylab=ylab, fig=f, color='orangered')
-plt.savefig('plots_'+remote_folder_path.name+'.png')
+plt.savefig('plots_'+data_folder_path.name+'.png')
 plt.show()
 
-# Saving results
+
+# create results folder
+#if not os.path.exists(str('analysis_'+ data_folder_path.name)):
+#    os.makedirs(str('analysis_'+ data_folder_path.name))
+    
 if not os.path.exists(str('analysis_'+ name_final_folder)):
-    os.makedirs(str('analysis_'+ name_final_folder))
+    os.makedirs(str('analysis_'+ name_final_folder))    
+
 #figure_path 
-pathlib.Path().absolute().joinpath('plots_'+ remote_folder_path.name +'.png').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'plots_'+ remote_folder_path.name +'.png'))
+pathlib.Path().absolute().joinpath('plots_'+ data_folder_path.name +'.png').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'plots_'+ data_folder_path.name +'.png'))
 #metadata_path
-pathlib.Path().absolute().joinpath('metadata_'+ remote_folder_path.name +'.txt').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'metadata_'+ remote_folder_path.name +'.txt'))
+pathlib.Path().absolute().joinpath('metadata_'+ data_folder_path.name +'.txt').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'metadata_'+ data_folder_path.name +'.txt'))
 #dataframe_path 
-pathlib.Path().absolute().joinpath('dataframe_' + remote_folder_path.name +'.csv').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'dataframe_'+ remote_folder_path.name +'.csv'))
+pathlib.Path().absolute().joinpath('dataframe_' + data_folder_path.name +'.csv').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'dataframe_'+ data_folder_path.name +'.csv'))
 #pdf_path 
-pathlib.Path().absolute().joinpath('pdf_report_' + remote_folder_path.name +'.pdf').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'pdf_report_'+ remote_folder_path.name +'.pdf'))
+pathlib.Path().absolute().joinpath('pdf_report_' + data_folder_path.name +'.pdf').rename(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder),'pdf_report_'+ data_folder_path.name +'.pdf'))
 # copy output file
 shutil.copyfile(pathlib.Path().absolute().joinpath(output_name),    pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder), output_name) )
 
-# making a zip file
-shutil.make_archive(str('analysis_'+ name_final_folder),'zip', pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)))
+#if send_data_to_NAS == 1:
+#  local_file_to_send_to_NAS = pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)+'.zip')
+#  fa.NASConnection(path_to_config_file,share_name = share_name).write_files_to_NAS(local_file_to_send_to_NAS, remote_folder_path)
 
-# Writing data to NAS
-if send_data_to_NAS == 1:
-  local_file_to_send_to_NAS = pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)+'.zip')
-  fa.NASConnection(path_to_config_file,share_name = share_name).write_files_to_NAS(local_file_to_send_to_NAS, remote_folder_path)
+# shutil.make_archive(str('analysis_'+ name_final_folder),'zip', pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)))
 
-#Moving all results to "analyses" folder
+# Writing analyses data to NAS
+if connect_to_NAS == True:
+    shutil.make_archive(str('analysis_'+ name_final_folder),'zip',pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)))
+    local_file_to_send_to_NAS = pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)+'.zip')
+    fa.NASConnection(path_to_config_file,share_name = share_name).write_files_to_NAS(local_file_to_send_to_NAS, data_folder_path)
+    os.remove(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)+'.zip'))
+    
+    
+# Writing masks to NAS
+if (connect_to_NAS == True) and (path_to_masks_dir == None) :
+    #shutil.make_archive(str('masks_'+ name_final_folder),'zip',pathlib.Path().absolute().joinpath(str('masks_'+ name_final_folder)))
+    shutil.make_archive(str('masks_'+ data_folder_path.name),'zip',pathlib.Path().absolute().joinpath(str('masks_'+ data_folder_path.name)))
+    #shutil.make_archive(str('masks_'+ data_folder_path.name),'zip',pathlib.Path().absolute().joinpath(str('masks_'+ name_final_folder)))
+    local_file_to_send_to_NAS = pathlib.Path().absolute().joinpath(str('masks_'+ data_folder_path.name)+'.zip')
+    fa.NASConnection(path_to_config_file,share_name = share_name).write_files_to_NAS(local_file_to_send_to_NAS, data_folder_path)
+    os.remove(pathlib.Path().absolute().joinpath(str('masks_'+ data_folder_path.name)+'.zip'))
+    
+    
+
+# Moving all results to "analyses" folder
 if not os.path.exists(str('analyses')):
     os.makedirs(str('analyses'))
-
+# Subfolder name
 analysis_folder_name = str('analysis_'+ name_final_folder)
 final_dir_name =pathlib.Path().absolute().joinpath('analyses', analysis_folder_name)
-
+# Removing directory if exist
 if os.path.exists(str(final_dir_name)):
   shutil.rmtree(str(final_dir_name))
-
+# Movng results to a subdirectory in 'analyses' folder
 pathlib.Path().absolute().joinpath(analysis_folder_name).rename(final_dir_name )
-# This line of code will replace the output file and place it on the specific directory. Notice that additional data is added after sending to NAS.
-shutil.move(pathlib.Path().absolute().joinpath(output_name),    final_dir_name.joinpath( output_name ) )
 
-# Delete local files
-shutil.rmtree(local_folder_path)
-temp_results_folder_name = pathlib.Path().absolute().joinpath('temp_results_' + remote_folder_path.name)
+# Delete local temporal files
+temp_results_folder_name = pathlib.Path().absolute().joinpath('temp_results_' + data_folder_path.name)
 shutil.rmtree(temp_results_folder_name)
-os.remove(pathlib.Path().absolute().joinpath(str('analysis_'+ name_final_folder)+'.zip'))
+
+
+
